feat: handle update failures
When starting, Peerio checks if the last update failed (this is designed
to happen when starting a new version right after the update), and if
so, it will show the error to the user and ask them to try the update
again, download the new version manually, or continue use the current
(older) version.

This also changes the "update is ready to install" dialog from alert()
to PeerUI Dialog and introduces another dialog to indicate that the
update is in progress.

Languages are now loaded before mounting the root component to avoid
race condition when something may have been showed before translations
are initialized.