feat: install updates on quit, ask to update every 12 hours
When the update is downloaded and it's not mandatory, it's automatically
scheduled to be installed when the user quits the app. If the user
doesn't quit within 12 hours, they are asked to install update (via a
dialog), with an option to postpone ("Later" button). The user will be
asked again in 12 hours, up to 2 times. The third time the dialog
appears, there's no "Later" option and the user will have to update.

If the update is mandatory, the user is asked to install it immediately
without an option to postpone.